added README file, environment.yml, requirements.txt, added code for creating portfoliio data based upon risk score
