+++ environment.yml
@@ -1,0 +1,130 @@
+name: project1
+channels:
+  - conda-forge
+  - defaults
+dependencies:
+  - aiofiles=22.1.0=pyhd8ed1ab_0
+  - aiosqlite=0.18.0=pyhd8ed1ab_0
+  - anyio=3.6.2=pyhd8ed1ab_0
+  - appnope=0.1.3=pyhd8ed1ab_0
+  - argon2-cffi=21.3.0=pyhd8ed1ab_0
+  - argon2-cffi-bindings=21.2.0=py311h5547dcb_3
+  - asttokens=2.2.1=pyhd8ed1ab_0
+  - attrs=22.2.0=pyh71513ae_0
+  - babel=2.11.0=pyhd8ed1ab_0
+  - backcall=0.2.0=pyh9f0ad1d_0
+  - backports=1.0=pyhd8ed1ab_3
+  - backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
+  - beautifulsoup4=4.11.2=pyha770c72_0
+  - bleach=6.0.0=pyhd8ed1ab_0
+  - brotlipy=0.7.0=py311h5547dcb_1005
+  - bzip2=1.0.8=h0d85af4_4
+  - ca-certificates=2022.12.7=h033912b_0
+  - certifi=2022.12.7=pyhd8ed1ab_0
+  - cffi=1.15.1=py311ha86e640_3
+  - charset-normalizer=2.1.1=pyhd8ed1ab_0
+  - comm=0.1.2=pyhd8ed1ab_0
+  - cryptography=39.0.1=py311h61927ef_0
+  - debugpy=1.6.6=py311h814d153_0
+  - decorator=5.1.1=pyhd8ed1ab_0
+  - defusedxml=0.7.1=pyhd8ed1ab_0
+  - entrypoints=0.4=pyhd8ed1ab_0
+  - executing=1.2.0=pyhd8ed1ab_0
+  - flit-core=3.8.0=pyhd8ed1ab_0
+  - idna=3.4=pyhd8ed1ab_0
+  - importlib-metadata=6.0.0=pyha770c72_0
+  - importlib_metadata=6.0.0=hd8ed1ab_0
+  - importlib_resources=5.10.2=pyhd8ed1ab_0
+  - ipykernel=6.21.1=pyh736e0ef_0
+  - ipython=8.10.0=pyhd1c38e8_0
+  - ipython_genutils=0.2.0=py_1
+  - jedi=0.18.2=pyhd8ed1ab_0
+  - jinja2=3.1.2=pyhd8ed1ab_1
+  - json5=0.9.5=pyh9f0ad1d_0
+  - jsonschema=4.17.3=pyhd8ed1ab_0
+  - jupyter_client=8.0.2=pyhd8ed1ab_0
+  - jupyter_core=5.2.0=py311h6eed73b_0
+  - jupyter_events=0.5.0=pyhd8ed1ab_1
+  - jupyter_server=2.2.1=pyhd8ed1ab_0
+  - jupyter_server_fileid=0.6.0=pyhd8ed1ab_0
+  - jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
+  - jupyter_server_ydoc=0.6.1=pyhd8ed1ab_0
+  - jupyter_ydoc=0.2.2=pyhd8ed1ab_0
+  - jupyterlab=3.6.1=pyhd8ed1ab_0
+  - jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
+  - jupyterlab_server=2.19.0=pyhd8ed1ab_0
+  - libcxx=14.0.6=hccf4f1f_0
+  - libffi=3.4.2=h0d85af4_5
+  - libsodium=1.0.18=hbcb3906_1
+  - libsqlite=3.40.0=ha978bb4_0
+  - libzlib=1.2.13=hfd90126_4
+  - markupsafe=2.1.2=py311h5547dcb_0
+  - matplotlib-inline=0.1.6=pyhd8ed1ab_0
+  - mistune=2.0.5=pyhd8ed1ab_0
+  - nbclassic=0.5.1=pyhd8ed1ab_0
+  - nbclient=0.7.2=pyhd8ed1ab_0
+  - nbconvert=7.2.9=pyhd8ed1ab_0
+  - nbconvert-core=7.2.9=pyhd8ed1ab_0
+  - nbconvert-pandoc=7.2.9=pyhd8ed1ab_0
+  - nbformat=5.7.3=pyhd8ed1ab_0
+  - ncurses=6.3=h96cf925_1
+  - nest-asyncio=1.5.6=pyhd8ed1ab_0
+  - notebook=6.5.2=pyha770c72_1
+  - notebook-shim=0.2.2=pyhd8ed1ab_0
+  - openssl=3.0.8=hfd90126_0
+  - packaging=23.0=pyhd8ed1ab_0
+  - pandoc=2.19.2=h694c41f_1
+  - pandocfilters=1.5.0=pyhd8ed1ab_0
+  - parso=0.8.3=pyhd8ed1ab_0
+  - pexpect=4.8.0=pyh1a96a4e_2
+  - pickleshare=0.7.5=py_1003
+  - pip=23.0=pyhd8ed1ab_0
+  - pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
+  - platformdirs=3.0.0=pyhd8ed1ab_0
+  - prometheus_client=0.16.0=pyhd8ed1ab_0
+  - prompt-toolkit=3.0.36=pyha770c72_0
+  - psutil=5.9.4=py311h5547dcb_0
+  - ptyprocess=0.7.0=pyhd3deb0d_0
+  - pure_eval=0.2.2=pyhd8ed1ab_0
+  - pycparser=2.21=pyhd8ed1ab_0
+  - pygments=2.14.0=pyhd8ed1ab_0
+  - pyopenssl=23.0.0=pyhd8ed1ab_0
+  - pyrsistent=0.19.3=py311h5547dcb_0
+  - pysocks=1.7.1=pyha2e5f31_6
+  - python=3.11.0=he7542f4_1_cpython
+  - python-dateutil=2.8.2=pyhd8ed1ab_0
+  - python-fastjsonschema=2.16.2=pyhd8ed1ab_0
+  - python-json-logger=2.0.4=pyhd8ed1ab_0
+  - python_abi=3.11=3_cp311
+  - pytz=2022.7.1=pyhd8ed1ab_0
+  - pyyaml=6.0=py311h5547dcb_5
+  - pyzmq=25.0.0=py311habfacb3_0
+  - readline=8.1.2=h3899abd_0
+  - requests=2.28.2=pyhd8ed1ab_0
+  - send2trash=1.8.0=pyhd8ed1ab_0
+  - setuptools=67.1.0=pyhd8ed1ab_0
+  - six=1.16.0=pyh6c4a22f_0
+  - sniffio=1.3.0=pyhd8ed1ab_0
+  - soupsieve=2.3.2.post1=pyhd8ed1ab_0
+  - stack_data=0.6.2=pyhd8ed1ab_0
+  - terminado=0.17.1=pyhd1c38e8_0
+  - tinycss2=1.2.1=pyhd8ed1ab_0
+  - tk=8.6.12=h5dbffcc_0
+  - tomli=2.0.1=pyhd8ed1ab_0
+  - tornado=6.2=py311h5547dcb_1
+  - traitlets=5.9.0=pyhd8ed1ab_0
+  - typing-extensions=4.4.0=hd8ed1ab_0
+  - typing_extensions=4.4.0=pyha770c72_0
+  - tzdata=2022g=h191b570_0
+  - urllib3=1.26.14=pyhd8ed1ab_0
+  - wcwidth=0.2.6=pyhd8ed1ab_0
+  - webencodings=0.5.1=py_1
+  - websocket-client=1.5.1=pyhd8ed1ab_0
+  - wheel=0.38.4=pyhd8ed1ab_0
+  - xz=5.2.6=h775f41a_0
+  - y-py=0.5.5=py311h890d03e_2
+  - yaml=0.2.5=h0d85af4_2
+  - ypy-websocket=0.8.2=pyhd8ed1ab_0
+  - zeromq=4.3.4=he49afe7_1
+  - zipp=3.13.0=pyhd8ed1ab_0
+
