+++ requirements.txt
@@ -1,0 +1,93 @@
+aiofiles==22.1.0
+aiosqlite==0.18.0
+anyio==3.6.2
+appnope==0.1.3
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+arrow==1.2.3
+asttokens==2.2.1
+attrs==22.2.0
+Babel==2.11.0
+backcall==0.2.0
+beautifulsoup4==4.11.2
+bleach==6.0.0
+certifi==2022.12.7
+cffi==1.15.1
+charset-normalizer==3.0.1
+comm==0.1.2
+debugpy==1.6.6
+decorator==5.1.1
+defusedxml==0.7.1
+executing==1.2.0
+fastjsonschema==2.16.2
+fqdn==1.5.1
+idna==3.4
+ipykernel==6.21.1
+ipython==8.10.0
+ipython-genutils==0.2.0
+isoduration==20.11.0
+jedi==0.18.2
+Jinja2==3.1.2
+json5==0.9.11
+jsonpointer==2.3
+jsonschema==4.17.3
+jupyter-events==0.5.0
+jupyter-ydoc==0.2.2
+jupyter_client==8.0.2
+jupyter_core==5.2.0
+jupyter_server==2.2.1
+jupyter_server_fileid==0.6.0
+jupyter_server_terminals==0.4.4
+jupyter_server_ydoc==0.6.1
+jupyterlab==3.6.1
+jupyterlab-pygments==0.2.2
+jupyterlab_server==2.19.0
+MarkupSafe==2.1.2
+matplotlib-inline==0.1.6
+mistune==2.0.5
+nbclassic==0.5.1
+nbclient==0.7.2
+nbconvert==7.2.9
+nbformat==5.7.3
+nest-asyncio==1.5.6
+notebook==6.5.2
+notebook_shim==0.2.2
+packaging==23.0
+pandocfilters==1.5.0
+parso==0.8.3
+pexpect==4.8.0
+pickleshare==0.7.5
+platformdirs==3.0.0
+prometheus-client==0.16.0
+prompt-toolkit==3.0.36
+psutil==5.9.4
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pycparser==2.21
+Pygments==2.14.0
+pyrsistent==0.19.3
+python-dateutil==2.8.2
+python-json-logger==2.0.5
+pytz==2022.7.1
+PyYAML==6.0
+pyzmq==25.0.0
+requests==2.28.2
+rfc3339-validator==0.1.4
+rfc3986-validator==0.1.1
+Send2Trash==1.8.0
+six==1.16.0
+sniffio==1.3.0
+soupsieve==2.3.2.post1
+stack-data==0.6.2
+terminado==0.17.1
+tinycss2==1.2.1
+tornado==6.2
+traitlets==5.9.0
+uri-template==1.2.0
+urllib3==1.26.14
+wcwidth==0.2.6
+webcolors==1.12
+webencodings==0.5.1
+websocket-client==1.5.1
+y-py==0.5.5
+ypy-websocket==0.8.2
